requirements.txt
# NEW-VOICE 2.0 - Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LLM Providers
groq==0.4.0              # Groq API (Llama 3.1) - MVP
openai==1.12.0           # OpenAI (fallback)
anthropic==0.18.0        # Claude (fallback)

# LiveKit
livekit==0.11.0
livekit-agents==0.7.0

# Speech-to-Text
deepgram-sdk==3.1.0

# Text-to-Speech
cartesia==1.0.0

# Vector Database
qdrant-client==1.7.0

# Database
supabase==2.3.0

# Utils
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1
pyyaml==6.0.1

# Audio processing
numpy==1.26.0
soundfile==0.12.1
